* Fix all linting errors * Move simulateEvent function to separate file
